# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: fn-cloudsync API
  version: 0.1.0
  description: Minimal control-plane and runtime API for the single-node WebDAV cloud sync service.
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/v1/tasks:
    get:
      summary: List tasks
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /api/v1/tasks/{taskID}:
    parameters:
      - $ref: "#/components/parameters/TaskID"
    get:
      summary: Get task
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/Error"
    put:
      summary: Update task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      summary: Delete task
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
...
    TaskID:
      name: taskID
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Task:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        connection_id:
          type: string
        local_path:
          type: string
        remote_path:
          type: string
        direction:
          type: string
        poll_interval_sec:
          type: integer
        conflict_policy:
          type: string
        delete_policy:
          type: string
        empty_dir_policy:
          type: string
        bandwidth_limit_kbps:
          type: integer
        max_workers:
          type: integer
        encryption_enabled:
          type: boolean
        hash_mode:
          type: string
        status:
          type: string
        desired_state:
          type: string
        last_error:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, connection_id, local_path, remote_path, direction, status, created_at, updated_at]
    CreateTaskRequest:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        connection_id:
          type: string
        local_path:
          type: string
        remote_path:
          type: string
        direction:
          type: string
        poll_interval_sec:
          type: integer
        conflict_policy:
          type: string
        delete_policy:
          type: string
        empty_dir_policy:
          type: string
        bandwidth_limit_kbps:
          type: integer
        max_workers:
          type: integer
        encryption_enabled:
          type: boolean
        hash_mode:
          type: string
      required: [id, name, connection_id, local_path, remote_path, direction]
    UpdateTaskRequest:
      type: object
      properties:
        name:
          type: string
        connection_id:
          type: string
        local_path:
          type: string
        remote_path:
          type: string
        direction:
          type: string
        poll_interval_sec:
          type: integer
        conflict_policy:
          type: string
        delete_policy:
          type: string
        empty_dir_policy:
          type: string
        bandwidth_limit_kbps:
          type: integer
        max_workers:
          type: integer
        encryption_enabled:
          type: boolean
        hash_mode:
          type: string
...